move isInternalFieldName to util
